Group No 9 ,
Group Members: Akhil Raveendran, Anoop Martin, Ranjana Ajayakumar
Project : Producer-Consumer Problem
Abstract:
This project focuses on  the  design of a programming solution to the bounded-buffer problem using the producer and consumer processes. One of the existing approach  for this problem was based on three semaphores: empty and full, which count the number of empty and full slots in the buffer, and mutex, which is a binary semaphore that protects the actual insertion or removal of items in the buffer. Binary Semaphores are implemented using busy waiting loops which waits for its entry in the critical section and spins in loop moreover binary semaphore allows only single thread. In addition to these, binary semaphore  implementation consumes more CPU cycles .In this project, the goal is to use standard counting semaphores for empty and full and a mutex lock, instead of  a binary semaphore, to represent mutex. The producer and consumer will run as separate threads and move items in synchronization with empty, full and mutex. We can solve this using Pthreads or the Windows API